category delete logic

+++ src/app/components/group-list/group-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { CategoriesService } from 'src/app/services/categories.service';
-import { Category } from 'src/interfaces/interfaces';
+import { ShoppingListService } from 'src/app/services/shopping-list.service';
+import { Category, Items } from 'src/interfaces/interfaces';
 
 @Component({
   selector: 'app-group-list',
@@ -9,8 +10,10 @@ import { Category } from 'src/interfaces/interfaces';
 })
 export class GroupListComponent implements OnInit {
   groups: Category[] =[];
+  counts!: number;
+  items:Items[]=[];
 
-  constructor(private groupService:CategoriesService) { }
+  constructor(private groupService:CategoriesService, private itemsService: ShoppingListService) { }
 
   ngOnInit(): void {
     this.getGroups();
@@ -28,10 +31,29 @@ export class GroupListComponent implements OnInit {
   }
 
   removeGroup(id:string){
-    this.groupService.deleteGroup(id).subscribe({
-      next:(res) =>{
-        alert('Category Deleted');
-        this.getGroups();
+    this.itemsService.getAllItems().subscribe({
+      next:(res) => {
+        this.groupService.getGroupById(id).subscribe({
+          next:(res2) => {
+            let groupFilter;
+            groupFilter = res.data.filter(i => {
+              return i.category == res2.data.category_name;
+            })
+
+            if(groupFilter.length == 0){
+              this.groupService.deleteGroup(id).subscribe({
+                  next:(res) =>{
+                    alert('Category Deleted');
+                    this.getGroups();
+                  }
+                })
+            }else{
+              alert(`${res2.data.category_name} category has ${groupFilter.length} item(s) \nRemove item(s) before proceeding.`)
+            }
+            console.log(groupFilter);
+            
+          }
+        })
       }
     })
   }
